requirements.txt
Flask==2.3.3
fpdf==1.7.2
Pillow==10.0.0
nbformat==5.9.2